Added a proper Environment class.
Old environment is renamed to TempEnvironment
